requirements.txt
streamlit>=1.36.0
requests==2.31.0
python-dotenv==1.0.1
google-generativeai>=0.8.3
numpy>=2.1.0
cachetools>=5.3.0
